* Update the BTagWeightCalculator in Heppy with a newer version of it * Remove the duplicate copy in CMGTools/MonoXAnalysis and use the Heppy new one in the friend tree producer
